+++ src/nuke_test_runner/bbox_checks/bbox_checker.py
