+++ src/nuke_test_runner/handler/configuration.py
@@ -16,7 +16,7 @@ import itertools
 import json
 from typing import TYPE_CHECKING
 
-from nuke_test_runner.runner import Runner, RunnerException
+from nuke_test_runner.handler.runner import Runner, RunnerException
 
 if TYPE_CHECKING:
     from pathlib import Path
