+++ tests/conftest.py
@@ -5,8 +5,7 @@ except ModuleNotFoundError:
 from pathlib import Path
 
 import pytest
-
-from nuke_test_runner.configuration import find_configuration
+from handler.configuration import find_configuration
 
 nuke_test = pytest.mark.skipif(not find_configuration(Path(Path.cwd())))
