+++ tests/src/nuke_test_runner/handler/test_configuration.py
@@ -8,9 +8,8 @@ from tempfile import TemporaryDirectory
 from unittest.mock import MagicMock, patch
 
 import pytest
-
-from nuke_test_runner.configuration import find_configuration, load_runners
-from nuke_test_runner.runner import Runner, RunnerException
+from nuke_test_runner.handler.configuration import find_configuration, load_runners
+from nuke_test_runner.handler.runner import Runner, RunnerException
 
 
 @pytest.fixture()
