+++ tests/src/nuke_test_runner/handler/test_runner.py
@@ -3,8 +3,7 @@ from pathlib import Path
 from unittest.mock import MagicMock, patch
 
 import pytest
-
-from nuke_test_runner.runner import Runner, RunnerException
+from nuke_test_runner.handler.runner import Runner, RunnerException
 
 
 @pytest.mark.parametrize(
