--- tests/test_setup.py
@@ -1,8 +1,0 @@
-"""Simple script to test the CICD workflow."""
-
-# TODO remove file when first real test is added.
-
-
-def test_workflow_works():
-    """This test should pass."""
-    assert True
